fix entities on domain

+++ lib/app/module/bank_statement/data/models/bank_transactions_model.dart
@@ -7,7 +7,7 @@ class BankTransactionsModel extends BankTransactions {
     String? identification,
     String? date,
     int? value,
-    int? balanceAfterOperation,
+    double? balanceAfterOperation,
     int? operation,
   }) : super(
             type: type,
